+++ relay-config/src/config.rs
@@ -243,6 +243,8 @@ pub struct OverridableConfig {
     pub public_key: Option<String>,
     /// Outcome source
     pub outcome_source: Option<String>,
+    /// shutdown timeout
+    pub shutdown_timeout: Option<String>,
 }
 
 /// The relay credentials
@@ -574,6 +576,22 @@ impl Default for Limits {
             max_alloc_size: None,
         }
     }
+}
+
+/// Controls traffic steering.
+#[derive(Debug, Default, Deserialize, Serialize)]
+#[serde(default)]
+pub struct Routing {
+    /// Accept and forward unknown Envelope items to the upstream.
+    ///
+    /// Forwarding unknown items should be enabled in most cases to allow proxying traffic for newer
+    /// SDK versions. The upstream in Sentry makes the final decision on which items are valid. If
+    /// this is disabled, just the unknown items are removed from Envelopes, and the rest is
+    /// processed as usual.
+    ///
+    /// Defaults to `true` for all Relay modes other than processing mode. In processing mode, this
+    /// is disabled by default since the item cannot be handled.
+    accept_unknown_items: Option<bool>,
 }
 
 /// Http content encoding for both incoming and outgoing web requests.
@@ -1236,6 +1254,8 @@ struct ConfigValues {
     #[serde(default)]
     logging: relay_log::LogConfig,
     #[serde(default)]
+    routing: Routing,
+    #[serde(default)]
     metrics: Metrics,
     #[serde(default)]
     sentry: relay_log::SentryConfig,
@@ -1432,6 +1452,13 @@ impl Config {
                     return Err(ConfigError::new(ConfigErrorKind::InvalidValue)
                         .field("incomplete credentials"));
                 }
+            }
+        }
+
+        let limits = &mut self.values.limits;
+        if let Some(shutdown_timeout) = overrides.shutdown_timeout {
+            if let Ok(shutdown_timeout) = shutdown_timeout.parse::<u64>() {
+                limits.shutdown_timeout = shutdown_timeout;
             }
         }
 
@@ -1951,6 +1978,12 @@ impl Config {
     /// Return the statically configured Relays.
     pub fn static_relays(&self) -> &HashMap<RelayId, RelayInfo> {
         &self.values.auth.static_relays
+    }
+
+    /// Returns `true` if unknown items should be accepted and forwarded.
+    pub fn accept_unknown_items(&self) -> bool {
+        let forward = self.values.routing.accept_unknown_items;
+        forward.unwrap_or_else(|| !self.processing_enabled())
     }
 }